Handle minStake and maxRank in Msgs

+++ x/ccv/provider/keeper/legacy_proposal_test.go
@@ -294,8 +294,8 @@ func TestHandleConsumerModificationProposal(t *testing.T) {
 	expectedValidatorSetCap := uint32(20)
 	expectedAllowlistedValidator := "cosmosvalcons1wpex7anfv3jhystyv3eq20r35a"
 	expectedDenylistedValidator := "cosmosvalcons1nx7n5uh0ztxsynn4sje6eyq2ud6rc6klc96w39"
-	expectedMinStake := 0
-	expectedMaxValidatorRank := 20
+	expectedMinStake := int64(0)
+	expectedMaxValidatorRank := int32(20)
 	proposal := providertypes.NewConsumerModificationProposal("title", "description", chainID,
 		expectedTopN,
 		expectedValidatorsPowerCap,
